+++ resources/js/pages/Admin/Sidebar_sections/Finance.vue
