+++ resources/js/pages/Admin/Sidebar_sections/UserManagement.vue
@@ -1,0 +1,7 @@
+
+<template>
+    <div>
+        <h3>آنالیزها</h3>
+        <p>اطلاعات مربوط به آنالیزها نمایش داده خواهد شد.</p>
+    </div>
+</template>
