+++ resources/js/pages/Admin/Sidebar_sections/ads_management.vue
