--- resources/js/pages/layout/menu_component/GiftCard.vue
@@ -1,19 +1,0 @@
-<template>
-  <v-list-item>
-    <v-list-item-title className="H_menu">GiftCard</v-list-item-title>
-  </v-list-item>
-</template>
-
-<script setup>
-</script>
-
-<style scoped>
-.H_menu {
-  font-size: 12px;
-  padding-left: 5px;
-  padding-right: 90px;
-  margin: 0;
-  cursor: pointer;
-  transition: background-color 0.3s, transform 0.3s;
-}
-</style>
