--- resources/js/pages/views/dashboards/dashboard_main/functional.vue
@@ -1,92 +1,0 @@
-
-<template>
-  <VCard class="main-card pa-4" outlined>
-    <div class="dashboard-header">
-      <p class="text-body-1 mb-0 fontsize">{{ translate('functional.todayPerformance') }}</p>
-    </div>
-    <div class="dashboard-cards">
-      <VCard v-for="(item, index) in performanceStats" :key="index" class="stat-card d-flex text-secondary2 align-center " flat>
-        <VIcon :color="item.color" size="24">{{ item.icon }}</VIcon>
-        <div class="ms-3">
-          <div class="text-h6 mb-1 fontsize">{{ item.stats }}</div>
-          <div class="text-body-6 text-h6 text-secondary2  fontsize">{{ t(item.title) }}</div>
-        </div>
-      </VCard>
-    </div>
-  </VCard>
-</template>
-
-<script setup>
-import { translate } from "../../../../store/languageStore";
-
-const performanceStats = [
-  { title: "functional.todayReservations", stats: '35', icon: 'mdi-calendar-check', color: 'primary' },
-  { title: "functional.dailyRevenue", stats: '1.5M', icon: 'mdi-cash', color: 'success' },
-  { title: "functional.newUsers", stats: '45', icon: 'mdi-account-plus', color: 'warning' },
-];
-</script>
-
-<style >
-@import '../../../../../../resources/css/@core/template/lib/@layouts/styles/cards.scss';
-
-.text-body-1{
-  font-size: 20px !important;
-  font-family: "2  Baran",serif  !important;
-  color: var(--text-card-color--groups);
-  font-weight: bold;
-}
-
-.main-card {
-  width: 100%;
-  max-width: 800px;
-  border-radius: 12px;
-  box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.1);
-  height: 100%;
-  background-color: var(--backgroundcards-color--groups);
-
-}
-
-.dashboard-header {
-  display: flex;
-  justify-content: flex-end;
-  margin-bottom: 16px;
-  background-color: var(--backgroundcards-color--groups);
-
-}
-
-.dashboard-cards {
-  padding-top: 20px;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 16px;
-}
-
-.stat-card {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  border-radius: 8px;
-  transition: box-shadow 0.3s ease;
-  padding: 16px;
-  background-color: var(--backgroundcards2-color--groups) !important;
-
-}
-.text-secondary2{
-  color: #161626;
-  font-family: "2  Baran",serif  !important;
-
-}
-
-.stat-card:hover {
-  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
-}
-
-.stat-card .text-h6 {
-  font-weight: 600;
-}
-
-.stat-card  {
-  color: #757575;
-}
-
-</style>
