+++ routes/instanceWeb/auth.php
@@ -3,13 +3,13 @@
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
-Route::get('/register', function () {
+Route::get('/RegisterUser', function () {
     return Inertia::render('RegisterUser');
 });
 
-Route::get('/login', function () {
+Route::get('/Login', function () {
     return Inertia::render('Login');
-})->name('login');
+})->name('Login');
 
 Route::get('/twoStep', function () {
     return Inertia::render('layout/menu_component/two-steps');
@@ -22,3 +22,6 @@ Route::get('/reset-password', function () {
 Route::get('/verify-email', function () {
     return Inertia::render('layout/menu_component/verify-email');
 });
+Route::get('/forgot-password', function () {
+    return Inertia::render('layout/menu_component/forgot-password');
+})->name('forgot-password');
